Refactor: split tokens per org (#971)
This PR relates to:
https://github.com/nl-design-system/documentatie/issues/2114

It will split up component token sets per organisation, making it easier
to copy and paste a token set of a specific component from an
organisation.

Component tokens affected:

- Breadcrumb Navigation
- Code
- Code Block
- Color Sample
- Data Badge
- Drawer
- Form Field Label
- Heading
- Link
- Mark
- Number Badge
- Page Number Badge
- Paragraph
- Progress List
- Skip Link

P.S. Changed the order of tokens for most of the tokensets while I was
at it.